# apps/web/src/lib/openapi.json
{
  "components": {
    "schemas": {
      "Body_inspect_pdf_api_inspect_pdf_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "The PDF to inspect.",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_inspect_pdf_api_inspect_pdf_post",
        "type": "object"
      },
...
      "Body_pdf_to_md_api_pdf_to_md_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "The PDF to convert.",
            "title": "File",
            "type": "string"
          },
          "options": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional JSON string. Example: `{\"page_break\": false, \"with_images\": false, \"front_matter\": true, \"lang\": \"en\"}`.",
            "title": "Options"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_pdf_to_md_api_pdf_to_md_post",
        "type": "object"
      },
      "ConvertStats": {
        "properties": {
          "bullets": {
            "default": 0,
            "title": "Bullets",
            "type": "integer"
          },
          "headings": {
            "default": 0,
            "title": "Headings",
            "type": "integer"
          },
          "tables": {
            "default": 0,
            "title": "Tables",
            "type": "integer"
          }
        },
        "title": "ConvertStats",
        "type": "object"
      },
      "FontUsage": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sample": {
            "title": "Sample",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "number"
          }
        },
        "required": [
          "name",
          "size",
          "count",
          "sample"
        ],
        "title": "FontUsage",
        "type": "object"
      },
...
      "FrontMatter": {
        "properties": {
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "FrontMatter",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InspectPdfResponse": {
        "properties": {
          "body_size_pt": {
            "title": "Body Size Pt",
            "type": "number"
          },
          "fonts": {
            "items": {
              "$ref": "#/components/schemas/FontUsage"
            },
            "title": "Fonts",
            "type": "array"
          },
          "heading_sizes_pt": {
            "items": {
              "type": "number"
            },
            "title": "Heading Sizes Pt",
            "type": "array"
          },
          "needs_ocr": {
            "title": "Needs Ocr",
            "type": "boolean"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "tagged": {
            "title": "Tagged",
            "type": "boolean"
          }
        },
        "required": [
          "pages",
          "body_size_pt",
          "heading_sizes_pt",
          "fonts",
          "tagged",
          "needs_ocr"
        ],
        "title": "InspectPdfResponse",
        "type": "object"
      },
      "PdfToMdResponse": {
        "properties": {
          "front_matter": {
            "$ref": "#/components/schemas/FrontMatter"
          },
          "md": {
            "title": "Md",
            "type": "string"
          },
          "ocr_applied": {
            "default": false,
            "title": "Ocr Applied",
            "type": "boolean"
          },
          "ocr_images_applied": {
            "default": false,
            "title": "Ocr Images Applied",
            "type": "boolean"
          },
          "stats": {
            "$ref": "#/components/schemas/ConvertStats"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "md"
        ],
        "title": "PdfToMdResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "md-bridge",
      "url": "https://github.com/vinicq/md-bridge"
    },
    "description": "\n**md-bridge** is a small, opinionated HTTP service that converts between PDF\nand Markdown using hand-written heuristics. No external calls, deterministic\noutput: same input, same output, every run.\n\n## What it does\n\n- **`POST /api/pdf-to-md`**: extract structured Markdown from a PDF.\n- **`POST /api/md-to-pdf`**: render a Markdown file to PDF through Chromium.\n- **`POST /api/inspect-pdf`**: return diagnostics about a PDF (fonts, sizes, tagged-PDF check).\n- **`GET  /api/health`**: liveness probe.\n\n## Try it out\n\nThe interactive docs you are reading right now (powered by Swagger UI) let\nyou call any endpoint with a real file. Click **Try it out**, attach a file,\npress **Execute**.\n\nA walkthrough with `curl` examples lives in\n[docs/API.md](https://github.com/vinicq/md-bridge/blob/main/docs/API.md).\n\n## Limits\n\n- Upload cap: **500 MB** per request.\n- No persistence: every file is processed in a temporary directory and removed\n  before the response is returned.\n- No OCR (v1): scanned PDFs need Tesseract before being submitted.\n",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "summary": "Deterministic, heuristic conversions between PDF and Markdown.",
    "title": "md-bridge API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/inspect-pdf": {
      "post": {
        "description": "Run cheap diagnostics on a PDF without converting it.\n\nUseful as a pre-flight check before `/api/pdf-to-md` so the UI can warn the\nuser about scanned PDFs (which need OCR first), show the inferred body font\nsize, detect PDF/UA tagging, and list the fonts in use.\n\n### Quick curl example\n\n```bash\ncurl -X POST http://localhost:8000/api/inspect-pdf \\\n  -F \"file=@whitepaper.pdf\"\n```\n\n### Response\n\n```json\n{\n  \"pages\": 4,\n  \"body_size_pt\": 11.0,\n  \"heading_sizes_pt\": [18.0, 14.0, 12.5],\n  \"fonts\": [{\"name\": \"InterRegular\", \"size\": 11.0, \"count\": 12048, \"sample\": \"Lorem ipsum...\"}],\n  \"tagged\": true,\n  \"needs_ocr\": false\n}\n```\n\n`needs_ocr=true` means PyMuPDF extracted very little text per page. The PDF\nis almost certainly scanned, so run Tesseract before submitting.",
        "operationId": "inspect_pdf_api_inspect_pdf_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_inspect_pdf_api_inspect_pdf_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "body_size_pt": 11.0,
                  "fonts": [
                    {
                      "count": 12048,
                      "name": "InterRegular",
                      "sample": "Lorem ipsum...",
                      "size": 11.0
                    },
                    {
                      "count": 320,
                      "name": "InterBold",
                      "sample": "Introduction",
                      "size": 18.0
                    }
                  ],
                  "heading_sizes_pt": [
                    18.0,
                    14.0,
                    12.5
                  ],
                  "needs_ocr": false,
                  "pages": 4,
                  "tagged": true
                },
                "schema": {
                  "$ref": "#/components/schemas/InspectPdfResponse"
                }
              }
            },
            "description": "Diagnostics about the uploaded PDF."
          },
          "400": {
            "description": "Upload is not a PDF."
          },
          "413": {
            "description": "Upload exceeds the 500 MB cap."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Inspect a PDF (fonts, sizes, tagged, OCR hint)",
        "tags": [
          "inspect"
        ]
      }
    },
...
    "/api/pdf-to-md": {
      "post": {
        "description": "Extract a structured Markdown document from a PDF.\n\nThe conversion is deterministic and heuristic. Headings are detected from\nfont size and the PDF outline, lists from bullet glyphs, tables from\nPyMuPDF's table finder.\n\n### Request\n\n`multipart/form-data` with two fields:\n\n| field | required | description |\n|---|---|---|\n| `file` | yes | A `.pdf` file (max 500 MB). |\n| `options` | no | JSON string with the keys below. |\n\n```json\n{\n  \"page_break\": false,\n  \"with_images\": false,\n  \"front_matter\": true,\n  \"lang\": \"en\"\n}\n```\n\n### Quick curl example\n\n```bash\ncurl -X POST http://localhost:8000/api/pdf-to-md \\\n  -F \"file=@whitepaper.pdf\" \\\n  -F 'options={\"front_matter\": true}'\n```\n\n### Error envelope\n\nEvery failure returns:\n\n```json\n{ \"error\": { \"code\": \"<machine_readable>\", \"message\": \"<human>\", \"detail\": <optional> } }\n```",
        "operationId": "pdf_to_md_api_pdf_to_md_post",
        "parameters": [
          {
            "description": "Convert even when the PDF has no extractable text layer. Without this, a scanned PDF returns 422 `ocr_required` instead of empty markdown. Use it to bypass a false positive (e.g. an image-only cover).",
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "description": "Convert even when the PDF has no extractable text layer. Without this, a scanned PDF returns 422 `ocr_required` instead of empty markdown. Use it to bypass a false positive (e.g. an image-only cover).",
              "title": "Force",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_pdf_to_md_api_pdf_to_md_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "front_matter": {
                    "author": "Author Name",
                    "date": "2026-04-12",
                    "pages": 4,
                    "source": "whitepaper.pdf",
                    "title": "Whitepaper"
                  },
                  "md": "---\\ntitle: \"Whitepaper\"\\npages: 4\\n---\\n\\n# Introduction\\n\\nFirst paragraph...",
                  "stats": {
                    "bullets": 14,
                    "headings": 6,
                    "tables": 1
                  },
                  "warnings": []
                },
                "schema": {
                  "$ref": "#/components/schemas/PdfToMdResponse"
                }
              }
            },
            "description": "Successful conversion."
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "wrong_file_type",
                    "message": "Expected a PDF file (.pdf) but got: notes.txt"
                  }
                }
              }
            },
            "description": "Upload is not a PDF."
          },
          "413": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "payload_too_large",
                    "message": "Upload exceeds 500 MB limit."
                  }
                }
              }
            },
            "description": "Upload exceeds the 500 MB cap."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "invalid_options",
                    "detail": [
                      {
                        "loc": [
                          "page_break"
                        ],
                        "msg": "Input should be a valid boolean"
                      }
                    ],
                    "message": "options payload failed validation"
                  }
                }
              }
            },
            "description": "Malformed `options` payload."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Convert a PDF to structured Markdown",
        "tags": [
          "convert"
        ]
      }
    },
...
  },
...
}